Update Components: remove useBreakpointValue
After deploying the application I noticed that when updating the page there was a wait for the display of the correct component for desktop or mobile. For example: Tooltip(desktop) and Popover(mobile) - for a brief moment no component is shown, only when client-side detection occurs that the correct component appears. For this example we don't have big problems but for others there may be layout shifts.